Separate HTML processing as a mixin
