+++ javascript/src/tests/test_parser.ts
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 
-import {parseFromString} from '../dom.js';
 import {Parser} from '../parser.js';
 
 describe('Parser.parse', () => {
@@ -42,121 +41,5 @@ describe('Parser.parse', () => {
     const parser = new Parser({});
     const result = parser.parse('');
     expect(result).toEqual([]);
-  });
-});
-
-describe('Parser.applyElement', () => {
-  const checkEqual = (
-    model: {[key: string]: {[key: string]: number}},
-    inputHTML: string,
-    expectedHTML: string
-  ) => {
-    const inputDOM = parseFromString(inputHTML);
-    const inputDocument = inputDOM.querySelector('p') as HTMLElement;
-    const parser = new Parser(model);
-    parser.applyElement(inputDocument);
-    const expectedDocument = parseFromString(expectedHTML);
-    const expectedElement = expectedDocument.querySelector('p') as HTMLElement;
-    expect(inputDocument.isEqualNode(expectedElement)).toBeTrue();
-  };
-
-  it('should insert WBR tags where the sentence should break.', () => {
-    const inputHTML = '<p>xyzabcabc</p>';
-    const expectedHTML = `
-    <p style="word-break: keep-all; overflow-wrap: anywhere;"
-    >xyz<wbr>abc<wbr>abc</p>`;
-    const model = {
-      UW4: {a: 1001}, // means "should separate right before 'a'".
-    };
-    checkEqual(model, inputHTML, expectedHTML);
-  });
-
-  it('should insert WBR tags even it overlaps with other HTML tags.', () => {
-    const inputHTML = '<p>xy<a href="#">zabca</a>bc</p>';
-    const expectedHTML = `<p style="word-break: keep-all; overflow-wrap: anywhere;"
-    >xy<a href="#">z<wbr>abc<wbr>a</a>bc</p>`;
-    const model = {
-      UW4: {a: 1001}, // means "should separate right before 'a'".
-    };
-    checkEqual(model, inputHTML, expectedHTML);
-  });
-});
-
-describe('Parser.translateHTMLString', () => {
-  const defaultModel = {
-    UW4: {a: 1001}, // means "should separate right before 'a'".
-  };
-  const checkEqual = (
-    model: {[key: string]: {[key: string]: number}},
-    inputHTML: string,
-    expectedHTML: string
-  ) => {
-    const parser = new Parser(model);
-    const result = parser.translateHTMLString(inputHTML);
-    const resultDocument = parseFromString(result);
-    const expectedDocument = parseFromString(expectedHTML);
-    expect(resultDocument.isEqualNode(expectedDocument)).toBeTrue();
-  };
-
-  it('should output a html string with a SPAN parent with proper style attributes.', () => {
-    const inputHTML = 'xyzabcd';
-    const expectedHTML = `
-    <span style="word-break: keep-all; overflow-wrap: anywhere;">xyz<wbr>abcd</span>`;
-    checkEqual(defaultModel, inputHTML, expectedHTML);
-  });
-
-  it('should not add a SPAN parent if the input already has one single parent.', () => {
-    const inputHTML = '<p class="foo" style="color: red">xyzabcd</p>';
-    const expectedHTML = `
-    <p class="foo"
-       style="color: red; word-break: keep-all; overflow-wrap: anywhere;"
-    >xyz<wbr>abcd</p>`;
-    checkEqual(defaultModel, inputHTML, expectedHTML);
-  });
-
-  it('should return a blank string if the input is blank.', () => {
-    const inputHTML = '';
-    const expectedHTML = '';
-    checkEqual({}, inputHTML, expectedHTML);
-  });
-
-  it('should pass script tags as-is.', () => {
-    const inputHTML = 'xyz<script>alert(1);</script>xyzabc';
-    const expectedHTML = `<span
-    style="word-break: keep-all; overflow-wrap: anywhere;"
-    >xyz<script>alert(1);</script>xyz<wbr>abc</span>`;
-    checkEqual(defaultModel, inputHTML, expectedHTML);
-  });
-
-  it('script tags on top should be discarded by the DOMParser.', () => {
-    const inputHTML = '<script>alert(1);</script>xyzabc';
-    const expectedHTML = `<span
-    style="word-break: keep-all; overflow-wrap: anywhere;"
-    >xyz<wbr>abc</span>`;
-    checkEqual(defaultModel, inputHTML, expectedHTML);
-  });
-
-  it('should skip some specific tags.', () => {
-    const inputHTML = 'xyz<code>abc</code>abc';
-    const expectedHTML = `<span
-    style="word-break: keep-all; overflow-wrap: anywhere;"
-    >xyz<code>abc</code><wbr>abc</span>`;
-    checkEqual(defaultModel, inputHTML, expectedHTML);
-  });
-
-  it('should not ruin attributes of child elements.', () => {
-    const inputHTML = 'xyza<a href="#" hidden>bc</a>abc';
-    const expectedHTML = `<span
-    style="word-break: keep-all; overflow-wrap: anywhere;"
-    >xyz<wbr>a<a href="#" hidden>bc</a><wbr>abc</span>`;
-    checkEqual(defaultModel, inputHTML, expectedHTML);
-  });
-
-  it('should work with emojis.', () => {
-    const inputHTML = 'xyza🇯🇵🇵🇹abc';
-    const expectedHTML = `<span
-    style="word-break: keep-all; overflow-wrap: anywhere;"
-    >xyz<wbr>a🇯🇵🇵🇹<wbr>abc</span>`;
-    checkEqual(defaultModel, inputHTML, expectedHTML);
   });
 });
